+++ RELEASE_NOTES.md
@@ -36,8 +36,11 @@ been tested on Linux. As of this release, the latest known version to work with 
 
 ### New user-visible features
 
+- (visualizer) Add Lr-Wpan NetDevices support to the Pyviz visualizer.
+
 ### Bugs fixed
 
+- (visualizer) !2636 - Fix overlapping labels and simulation stop time in PyViz visualizer.
 - (wifi) Fix incorrect aPSDUMaxLength value for 802.11be.
 - (wifi) Fix hardcoded threshold value in EHT PHY to determine per-20MHz CCA indication.